+++ rest/src/index.ts
@@ -1,5 +1,8 @@
 export * from './decorators';
 export * from './middlewares';
 
+export * from './misc/defaults';
+export type {middlewareTypes} from './misc/middlewareTypes';
 export * from './misc/restHttpError.interface';
+export * from './misc/tokens';
 export * from './misc/withRestClientContext';